Algorithmus
Generelle Änderungen

+++ BlackmanNutall.h
@@ -2,6 +2,6 @@
 #define BLACKMANNUTALL_H_INCLUDED
 
 // Optimal overlap: 66.3%
-float* BlackmanNutall(int windowSize);
+double* BlackmanNutall(int windowSize);
 
 #endif // BLACKMANNUTALL_H_INCLUDED
